Refactor code structure + fix pip package issue (#9)
