+++ src/lab_checker/checks/AbstractCheck.py
