+++ src/lab_checker/checks/CheckResult.py
@@ -1,6 +1,6 @@
 import logging
 
-from kathara_lab_checker.utils import green, red
+from lab_checker.utils import green, red
 
 
 class CheckResult:
