+++ src/lab_checker/checks/CollisionDomainCheck.py
