+++ src/lab_checker/checks/CustomCommandCheck.py
