+++ src/lab_checker/checks/DaemonCheck.py
@@ -2,7 +2,7 @@ from Kathara.exceptions import MachineNotFoundError
 from Kathara.manager.Kathara import Kathara
 from Kathara.model.Lab import Lab
 
-from kathara_lab_checker.utils import get_output
+from lab_checker.utils import get_output
 from .AbstractCheck import AbstractCheck
 from .CheckResult import CheckResult
